+++ RELEASE_NOTES.md
@@ -6,6 +6,8 @@ ENHANCEMENTS:
 
 1. Message filter option added to ubxdump command line utility. See README for usage.
 
+2. MON-SPAN spectrum attribute now parsed as a single array of integers `spectrumRf`rather than as 256 separate integers `spectrum_01_01`, `spectrum_01_02`etc. for ease of processing and charting.
+
 ### RELEASE 1.0.15
 
 ENHANCEMENTS: